test: split the #8067 shape-authority tests out of parent_static.rs
parent_static.rs was at 1992 lines on main and this PR adds 52, crossing the
2000-line cap. Extracts the inline shape_authority_tests_8067 module to a
sibling under parent_static/; body unchanged.
